+++ History.md
@@ -5,3 +5,7 @@ Basic TCP Server and Protocol.
 ## Alpha 0.0.2
 In-Memery UUID to IP map.  
 A test File is added to test connections form 'client'.
+
+## Alpha 0.0.3
+Message Status Management and Forwarding.
+The Test file is updated to test message status management and forwarding.